refactor(LottoController, App): 로또 발매기 주요 로직 LottoController로 이동
